+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## 8.0.0
 
 ### Enhancements
+- **Survey Changes**
+  - Longitudinal surveys can be launched and assigned to users, ex:
+    - `s = Survey.find_by_slug 'about-me'`
+    - `s.launch_single(user, '6month')`
+    - or
+    - `s.launch_multiple(users, '6month')`
 - **Gem Changes**
   - Only run `web-console` in development mode